# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KropDiagnost API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "KropDiagnost API server"
        }
    ],
    "paths": {
        "/get-equipment/category-by-type/{type}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Return list of all equipments",
                "operationId": "App\\Http\\Controllers\\Api\\EquipmentCategoryByTypeController::__invoke",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Електрокардіографія"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "qty": {
                                                        "type": "integer"
                                                    },
                                                    "equipments": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "custodian_name": {
                                                                    "type": "string"
                                                                },
                                                                "custodian_identifier": {
                                                                    "type": "string"
                                                                },
                                                                "address_country_name": {
                                                                    "type": "string"
                                                                },
                                                                "address_region": {
                                                                    "type": "string"
                                                                },
                                                                "address_locality": {
                                                                    "type": "string"
                                                                },
                                                                "address_street_address": {
                                                                    "type": "string"
                                                                },
                                                                "latitude": {
                                                                    "type": "string"
                                                                },
                                                                "longitude": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_title": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_model": {
                                                                    "type": "string"
                                                                },
                                                                "sn_equipment": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_identifier": {
                                                                    "type": "string"
                                                                },
                                                                "producer_country": {
                                                                    "type": "string"
                                                                },
                                                                "producer_name": {
                                                                    "type": "string"
                                                                },
                                                                "producer_identifier": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_year": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_life": {
                                                                    "type": "integer"
                                                                },
                                                                "registration_date": {
                                                                    "type": "string"
                                                                },
                                                                "equip_condition": {
                                                                    "type": "string"
                                                                },
                                                                "repair_date": {
                                                                    "type": "string"
                                                                },
                                                                "repair_type": {
                                                                    "type": "string"
                                                                },
                                                                "structure_name": {
                                                                    "type": "string"
                                                                },
                                                                "floor_number": {
                                                                    "type": "integer"
                                                                },
                                                                "room_number": {
                                                                    "type": "integer"
                                                                },
                                                                "inspection_type": {
                                                                    "type": "string"
                                                                },
                                                                "diagnostic_subgroup": {
                                                                    "type": "string"
                                                                },
                                                                "diagnostic_type": {
                                                                    "type": "string"
                                                                },
                                                                "diagnostic_quantity": {
                                                                    "type": "integer"
                                                                },
                                                                "work_shedule": {
                                                                    "type": "string"
                                                                },
                                                                "availability_restriction": {
                                                                    "type": "string"
                                                                },
                                                                "created_at": {
                                                                    "type": "string"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string"
                                                                },
                                                                "id_u": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/hospital/{id_u}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Return hospital data by id_u",
                "operationId": "App\\Http\\Controllers\\Api\\HospitalController::__invoke",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_u"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "uniqueItems": true,
                                        "example": "fb78ab4f49574da62907afa386a404d4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "hospitals": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "custodian_name": {
                                                                    "type": "string"
                                                                },
                                                                "custodian_identifier": {
                                                                    "type": "string"
                                                                },
                                                                "address_country_name": {
                                                                    "type": "string"
                                                                },
                                                                "address_region": {
                                                                    "type": "string"
                                                                },
                                                                "address_locality": {
                                                                    "type": "string"
                                                                },
                                                                "address_street_address": {
                                                                    "type": "string"
                                                                },
                                                                "latitude": {
                                                                    "type": "string"
                                                                },
                                                                "longitude": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_title": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_model": {
                                                                    "type": "string"
                                                                },
                                                                "sn_equipment": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_identifier": {
                                                                    "type": "string"
                                                                },
                                                                "producer_country": {
                                                                    "type": "string"
                                                                },
                                                                "producer_name": {
                                                                    "type": "string"
                                                                },
                                                                "producer_identifier": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_year": {
                                                                    "type": "string"
                                                                },
                                                                "equipment_life": {
                                                                    "type": "integer"
                                                                },
                                                                "registration_date": {
                                                                    "type": "string"
                                                                },
                                                                "equip_condition": {
                                                                    "type": "string"
                                                                },
                                                                "repair_date": {
                                                                    "type": "string"
                                                                },
                                                                "repair_type": {
                                                                    "type": "string"
                                                                },
                                                                "structure_name": {
                                                                    "type": "string"
                                                                },
                                                                "floor_number": {
                                                                    "type": "integer"
                                                                },
                                                                "room_number": {
                                                                    "type": "integer"
                                                                },
                                                                "inspection_type": {
                                                                    "type": "string"
                                                                },
                                                                "diagnostic_subgroup": {
                                                                    "type": "string"
                                                                },
                                                                "diagnostic_type": {
                                                                    "type": "string"
                                                                },
                                                                "diagnostic_quantity": {
                                                                    "type": "integer"
                                                                },
                                                                "work_shedule": {
                                                                    "type": "string"
                                                                },
                                                                "availability_restriction": {
                                                                    "type": "string"
                                                                },
                                                                "created_at": {
                                                                    "type": "string"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string"
                                                                },
                                                                "id_u": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}